Add dynamic option to mapChannels
Adding dynamic option to signify that the channel name should be pulled from a property instead of the defined channel name.  Helps if you don't know what the channel name is ahead of time, e.g. when you have a component that could be instantiated in multiple places, but shouldn't clobber data for unrelated views